fix: Added Global Exception Handler in Author

+++ src/main/java/pk/edu/niit/library_management_system/Author/Services/AuthorServices.java
@@ -1,10 +1,12 @@
 package pk.edu.niit.library_management_system.Author.Services;
 
+import ch.qos.logback.classic.spi.IThrowableProxy;
 import lombok.Data;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pk.edu.niit.library_management_system.Author.Entity.Author;
 import pk.edu.niit.library_management_system.Author.Repository.AuthorRepository;
+import pk.edu.niit.library_management_system.ExceptionHandler.AuthorNotFoundException;
 
 import java.util.List;
 import java.util.Optional;
@@ -18,42 +20,42 @@ public class AuthorServices {
     public List<Author> getAllAuthors()
     {
         return authorRepository.findAll();
+
     }
 
     public Author createAuthor(Author author)
     {
+
         return authorRepository.save(author);
     }
 
 
     public void deleteAuthor(long id)
     {
+        if(!authorRepository.existsById(id))
+        {
+            throw new AuthorNotFoundException(
+                    "DELETE/author/id/{ "+id +"} :Author with this id " + id + "not found"
+            );
+        }
         authorRepository.deleteById(id);
     }
 
     public Author updateAuthor(long id,Author author)
     {
-        Optional<Author> existingAuthor=authorRepository.findById(id);
-        if(existingAuthor.isPresent())
-        {
-            Author updated=existingAuthor.get();
-            updated.setAuthorName(author.getAuthorName());
-            updated.setBio(author.getBio());
-            return authorRepository.save(updated);
-        }
+        Author existingAuthor=authorRepository.findById(id).orElseThrow(()-> new AuthorNotFoundException("PUT/author/id/: Author not found for this id: "+id));
 
-        return null;
+            existingAuthor.setAuthorName(author.getAuthorName());
+            existingAuthor.setBio(author.getBio());
+            return authorRepository.save(existingAuthor);
+
+
     }
 
     public Author getAuthorById(long id)
     {
-        Optional<Author> existingAuthor=authorRepository.findById(id);
-        if(existingAuthor.isPresent())
-        {
-            Author authorFound=existingAuthor.get();
+        return authorRepository.findById(id).orElseThrow(()->
+       new AuthorNotFoundException("Author Not Found with this " + id));
 
-            return authorFound;
-        }
-        return null;
     }
 }
